Delete shared GeraLanding backend client class
